+++ crates/starknet_api/src/state_test.rs
@@ -1,9 +1,6 @@
-use assert_matches::assert_matches;
-
-use crate::state::PatriciaKey;
 use crate::{
     shash, ClassHash, ContractAddress, ContractClass, ContractNonce, DeclaredContract,
-    DeployedContract, Nonce, StarkHash, StarknetApiError, StateDiff, StorageDiff,
+    DeployedContract, Nonce, StarkHash, StateDiff, StorageDiff,
 };
 
 #[test]
@@ -60,19 +57,4 @@ fn state_sorted() {
         state_diff.destruct(),
         (sorted_deployed_contracts, sorted_storage_diffs, sorted_declared_contracts, sorted_nonces)
     );
-}
-
-#[test]
-fn patricia_key_valid() {
-    let hash = shash!("0x123");
-    let patricia_key = PatriciaKey::new(hash).unwrap();
-    assert_eq!(patricia_key.0, hash);
-}
-
-#[test]
-fn patricia_key_out_of_range() {
-    // 2**251
-    let hash = shash!("0x800000000000000000000000000000000000000000000000000000000000000");
-    let err = PatriciaKey::new(hash);
-    assert_matches!(err, Err(StarknetApiError::OutOfRange { string: _err_str }));
 }
